pyproject.toml
[project]
name = "pixel-dialogue"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "accelerate>=1.12.0",
    "diffusers>=0.36.0",
    "easyocr>=1.7.2",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.2.1",
    "langchain-groq>=1.1.2",
    "langchain-openai>=1.1.10",
    "langchain-tavily>=0.2.17",
    "langgraph>=1.0.10",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.2.2",
    "streamlit>=1.54.0",
    "torch>=2.6.0",
    "transformers>=5.2.0",
    "ultralytics>=8.4.19",
]

[[tool.uv.index]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu124"
explicit = true

[tool.uv.sources]
torch = [{ index = "pytorch" }]
torchvision = [{ index = "pytorch" }]
torchaudio = [{ index = "pytorch" }]
